requirements.txt
asgiref==3.11.1
Django==6.0.3
django-unfold==0.83.1
pillow==12.1.1
python-decouple==3.8
sqlparse==0.5.5
